Add support for sponsors.
This is what a commit to add a new module will generally look like.
It includes a basic framework for setting up alerts; a support function
that will make API calls from the background runner; a set of models for
storing the configured alerts; and support calls for converting those
models to alerts.

Outside of the app, you'll need to configure the URLs, app, and the
background runner to have support for your new module.