Add a new column for next update.
This allows for exponential backoff in case of failure, and for showing
users when last *successful* fetch was, useful for debugging failures
after the fact. (last_update no longer updated on failure.) Also, add
a default time to both fields, and remove it from downstream views that
were all using the same value anyway.