Add support for loading any Google font.
Stash as list of the current Google fonts, and add a boolean property to alerts
which lists whether the font is a Google font based on string check.
When loading a font which is a Google font, use the webfont loader to request
the font. Add a preloader to load recently used fonts (should be extended to
apply to images and sounds as well).