Remove default SECRET_KEY variables.
When running a Django app, you should always set your own SECRET_KEY variable
in the app settings. Including a default SECRET_KEY in the settings means
some folks might not realize they should change this, so replacing it with
CHANGEME to make it more clear. (Did the same for the database password,
which should obviously also not be "password".)